hailuo/pyproject.toml
[project]
name = "hailuo-cli"
version = "0.1.0"
description = "CLI tool for Hailuo AI Video Generation via AceDataCloud API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "AceDataCloud", email = "support@acedata.cloud" }
]
maintainers = [
    { name = "AceDataCloud", email = "support@acedata.cloud" }
]
keywords = [
    "cli",
    "hailuo",
    "minimax",
    "ai",
    "video",
    "generation",
    "acedata",
    "command-line",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: Console",
]
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
]
release = [
    "build>=1.2.0",
    "twine>=6.1.0",
]
all = [
    "hailuo-cli[dev,test,release]",
]

[project.scripts]
hailuo-cli = "hailuo_cli.main:cli"
hailuo = "hailuo_cli.main:cli"

[project.urls]
Homepage = "https://github.com/AceDataCloud/HailuoCli"
Repository = "https://github.com/AceDataCloud/HailuoCli"
Issues = "https://github.com/AceDataCloud/HailuoCli/issues"
Changelog = "https://github.com/AceDataCloud/HailuoCli/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling>=1.21.0,<1.22.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hailuo_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "hailuo_cli/",
    "tests/",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    ".env.example",
]

# Mypy Configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require real API access",
    "slow: marks slow tests",
]

# Ruff Configuration
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
